Renamed PDFDocument to GiniPDFDocument and moved to a separated file.
Issue #50
